+++ lib/view_component/config.rb
@@ -115,9 +115,6 @@ module ViewComponent
       # The locations in which component previews will be looked up.
       # Defaults to `['test/component/previews']` relative to your Rails root.
 
-      # @!attribute preview_path
-      # @deprecated Use #preview_paths instead. Will be removed in v3.0.0.
-
       # @!attribute test_controller
       # @return [String]
       # The controller used for testing components.
@@ -139,15 +136,6 @@ module ViewComponent
 
     def initialize
       @config = self.class.defaults
-    end
-
-    def preview_path
-      preview_paths
-    end
-
-    def preview_path=(new_value)
-      ViewComponent::Deprecation.warn("`preview_path` will be removed in v3.0.0. Use `preview_paths` instead.")
-      self.preview_paths = Array.wrap(new_value)
     end
 
     delegate_missing_to :config