+++ lib/view_component/test_helpers.rb
@@ -31,18 +31,6 @@ module ViewComponent
     # @private
     attr_reader :rendered_content
 
-    # Returns the result of a render_inline call.
-    #
-    # @return [String]
-    def rendered_component
-      ViewComponent::Deprecation.warn(
-        "`rendered_component` is deprecated and will be removed in v3.0.0. " \
-        "Use `page` instead."
-      )
-
-      rendered_content
-    end
-
     # Render a component inline. Internally sets `page` to be a `Capybara::Node::Simple`,
     # allowing for Capybara assertions to be used:
     #