+++ test/sandbox/test/rendering_test.rb
@@ -25,12 +25,6 @@ class RenderingTest < ViewComponent::TestCase
     assert_includes rendered_content, "hello,world!"
   end
 
-  def test_render_inline_sets_rendered_component
-    render_inline(MyComponent.new)
-
-    assert_includes rendered_component, "hello,world!"
-  end
-
   def test_child_component
     render_inline(ChildComponent.new)