+++ docs/CHANGELOG.md
@@ -10,7 +10,7 @@ nav_order: 5
 
 ## main
 
-* BREAKING: Remove deprecated support for loading ViewComponent engine manually. Make sure you removed `require "view_component/engine"` from your `Gemfile`.
+* BREAKING: Remove deprecated support for loading ViewComponent engine manually. Make sure `require "view_component/engine"` is removed from `Gemfile`.
 
     *Joel Hawksley*